src/scrapers/lesoir/requirements.txt
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
comm==0.1.4
decorator==5.1.1
dnspython==2.4.1
executing==1.2.0
idna==3.4
lxml==4.9.3
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
nest-asyncio==1.5.7
numpy==1.25.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
stack-data==0.6.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
